+++ src/Model/Custom.hs
@@ -1,31 +1,30 @@
-
 module Model.Custom where
 
-import Prelude
-
 import Crypto.BCrypt as Import hiding (hashPassword)
-import Database.Persist.Sql
-import Safe (fromJustNote)
-import qualified Data.Text as T
-import qualified Data.Text.Encoding as TE
+import qualified Crypto.Hash.SHA256 as SHA256
 import qualified Data.Aeson as A
-import System.Entropy (getEntropy)
-import qualified Data.ByteString.Builder as BB
-import qualified Data.ByteString.Lazy as LBS
 import qualified Data.Base64.Types as Base64
 import qualified Data.ByteString.Base64.URL as Base64Url
-import qualified Crypto.Hash.SHA256 as SHA256
+import qualified Data.ByteString.Builder as BB
+import qualified Data.ByteString.Lazy as LBS
+import qualified Data.Text as T
+import qualified Data.Text.Encoding as TE
+import Database.Persist.Sql
+import Safe (fromJustNote)
+import System.Entropy (getEntropy)
+import Prelude
 
 mkSlug :: Int -> IO T.Text
 mkSlug size =
-  TE.decodeUtf8 . LBS.toStrict . BB.toLazyByteString . BB.byteStringHex <$>
-  getEntropy size
+  TE.decodeUtf8 . LBS.toStrict . BB.toLazyByteString . BB.byteStringHex
+    <$> getEntropy size
 
 -- * Bookmark Slug
 
 newtype BmSlug = BmSlug
   { unBmSlug :: T.Text
-  } deriving (Eq, PersistField, PersistFieldSql, Show, Read, Ord, A.FromJSON, A.ToJSON)
+  }
+  deriving (Eq, PersistField, PersistFieldSql, Show, Read, Ord, A.FromJSON, A.ToJSON)
 
 mkBmSlug :: IO BmSlug
 mkBmSlug = BmSlug <$> mkSlug 6
@@ -34,7 +33,8 @@ mkBmSlug = BmSlug <$> mkSlug 6
 
 newtype NtSlug = NtSlug
   { unNtSlug :: T.Text
-  } deriving (Eq, PersistField, PersistFieldSql, Show, Read, Ord, A.FromJSON, A.ToJSON)
+  }
+  deriving (Eq, PersistField, PersistFieldSql, Show, Read, Ord, A.FromJSON, A.ToJSON)
 
 mkNtSlug :: IO NtSlug
 mkNtSlug = NtSlug <$> mkSlug 10
@@ -44,13 +44,14 @@ mkNtSlug = NtSlug <$> mkSlug 10
 policy :: HashingPolicy
 policy =
   HashingPolicy
-  { preferredHashCost = 12
-  , preferredHashAlgorithm = "$2a$"
-  }
+    { preferredHashCost = 12,
+      preferredHashAlgorithm = "$2a$"
+    }
 
 newtype BCrypt = BCrypt
   { unBCrypt :: T.Text
-  } deriving (Eq, PersistField, PersistFieldSql, Show, Ord, A.FromJSON, A.ToJSON)
+  }
+  deriving (Eq, PersistField, PersistFieldSql, Show, Ord, A.FromJSON, A.ToJSON)
 
 hashPassword :: T.Text -> IO BCrypt
 hashPassword rawPassword = do
@@ -62,7 +63,7 @@ validatePasswordHash :: BCrypt -> T.Text -> Bool
 validatePasswordHash hash' pass = do
   validatePassword (TE.encodeUtf8 (unBCrypt hash')) (TE.encodeUtf8 pass)
 
-newtype ApiKey = ApiKey { unApiKey :: T.Text }
+newtype ApiKey = ApiKey {unApiKey :: T.Text}
 
 newtype HashedApiKey
   = HashedApiKey T.Text